Todo funcionando
La demo ya funciona con todo.

+++ proyecto/proyecto/src/main/java/GI/proyecto/controller/webController.java
@@ -1,6 +1,7 @@
 package GI.proyecto.controller;
 
 import java.util.List;
+import java.util.Optional;
 
 import GI.proyecto.model.tSolucion;
 import GI.proyecto.model.tUsuario;
@@ -9,12 +10,16 @@ import GI.proyecto.service.tUsuarioService;
 import jakarta.servlet.http.HttpSession;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.servlet.ModelAndView;
 
 import GI.proyecto.model.tMuestra;
+import GI.proyecto.model.tPermiso;
 import GI.proyecto.service.tMuestraService;
+import GI.proyecto.service.tPermisoService;
 
 @Controller
 public class webController {
@@ -27,13 +32,18 @@ public class webController {
 	@Autowired
 	tUsuarioService usuarioService;
 
+	@Autowired
+	tPermisoService permisoService;
+
 	@GetMapping("/muestra")
 	public String listarMuestras(Model model, HttpSession sesion) {
 		tMuestra muestra = new tMuestra();
 		tUsuario usuario = usuarioService.findById((String) sesion.getAttribute("nif")).get(0); // sacamos de la sesion
 																								// el nif pasada en el
 																								// login
+		tPermiso permiso = permisoService.findPermiso(usuario.getRolName());
 		model.addAttribute("usuario", usuario);
+		model.addAttribute("permiso", permiso);
 		model.addAttribute("muestra", muestra);
 
 		List<tMuestra> muestrasList = muestraService.getAll();
@@ -53,6 +63,8 @@ public class webController {
 	@RequestMapping("/mostrarMuestra")
 	public String doShowMostrar(Model model, @RequestParam("muestraId") Integer muestraId, HttpSession sesion) {
 		tUsuario usuario = usuarioService.findById((String) sesion.getAttribute("nif")).get(0);
+		tPermiso permiso = permisoService.findPermiso(usuario.getRolName());
+		model.addAttribute("permiso", permiso);
 		model.addAttribute("usuario", usuario);
 
 		tMuestra muestra = muestraService.getMuestra(muestraId);
@@ -79,11 +91,18 @@ public class webController {
 	public String postLogin(tUsuario usuario, HttpSession sesion) {
 
 		List<tUsuario> user = usuarioService.findById(usuario.getNif());
-		if (user.size() == 1 && user.get(0).getPassword().equals(usuario.getPassword())) { // podemos usar el bCrypt
 
-			sesion.setAttribute("nif", user.get(0).getNif());// Al hacer login metemos por la sesión el
-																// nif para poder hacer uso del usuario en otras vistas
-			return "redirect:/muestra";
+		if (user.size() == 1 && user.get(0).getPassword().equals(usuario.getPassword())) { // podemos usar el bCrypt
+			tPermiso permiso = permisoService.findPermiso(user.get(0).getRolName());
+			if (permiso.getAcceso()) {
+				sesion.setAttribute("nif", user.get(0).getNif()); // Al hacer login metemos por la sesión el
+																	// nif para poder hacer uso del usuario en otras
+																	// vistas
+				return "redirect:/muestra";
+			}else{
+				return "error";
+			}
+
 		} else {
 			return "error";
 		}
@@ -106,9 +125,14 @@ public class webController {
 	}
 
 	@RequestMapping("/borrar-muestra/{id}")
-	public String cargarPaginaBorrar(@PathVariable Integer id) {
-		muestraService.delete(id);
-		return "redirect:/muestra";
+	public String cargarPaginaBorrar(@PathVariable Optional<Integer> optionalId) {
+		if(optionalId.isPresent()) {
+			Integer id = optionalId.get();
+			muestraService.delete(id);
+			return "redirect:/muestra";
+
+		}
+		return "/";
 
 	}
 
@@ -120,5 +144,13 @@ public class webController {
 
 		return "redirect:/muestra";
 	}
+	
+	@ExceptionHandler(Exception.class)
+    @ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR)
+    public ModelAndView handleException(Exception e) {
+        ModelAndView modelAndView = new ModelAndView("error");
+        modelAndView.addObject("errorDetails", e.getMessage());
+        return modelAndView;
+    }
 
 }
